// streampark-console/streampark-console-webapp-v2/src/types/api/flink/app.type.ts
export interface AppListRecord {
    id: string
    jobType: number
    projectId?: any
    tags?: any
    userId: string
    teamId: string
    jobName: string
    appId?: string
    jobId?: string
    versionId: string
    clusterId?: string
    flinkImage?: string
    k8sNamespace: string
    serviceAccount?: string
    state: number
    release: number
    build: boolean
    restartSize: number
    restartCount?: number
    optionState: number
    alertId?: any
    args?: string
    module?: any
    options: string
    hotParams?: string
    resolveOrder: number
    deployMode: number
    dynamicProperties?: string
    appType: number
    tracking: number
    jar?: string
    jarCheckSum?: string
    mainClass?: string
    startTime?: string
    endTime?: string
    duration?: string
    cpMaxFailureInterval?: any
    cpFailureRateInterval?: any
    cpFailureAction?: any
    totalTM?: any
    totalSlot?: any
    availableSlot?: any
    jmMemory?: number
    tmMemory?: number
    totalTask?: number
    flinkClusterId?: any
    description?: string
    createTime: string
    optionTime?: string
    modifyTime: string
    k8sRestExposedType?: any
    k8sPodTemplate?: any
    k8sJmPodTemplate?: any
    k8sTmPodTemplate?: any
    ingressTemplate?: any
    defaultModeIngress?: any
    resourceFrom?: number
    k8sHadoopIntegration: boolean
    overview?: any
    teamResource?: any
    dependency?: any
    sqlId?: any
    flinkSql?: any
    stateArray?: any
    jobTypeArray?: any
    backUp: boolean
    restart: boolean
    userName: string
    nickName: string
    config?: any
    configId?: any
    flinkVersion: string
    confPath?: any
    format?: any
    savepointPath?: any
    restoreOrTriggerSavepoint: boolean
    drain: boolean
    nativeFormat: boolean
    allowNonRestored: boolean
    socketId?: any
    projectName?: any
    createTimeFrom?: any
    createTimeTo?: any
    backUpDescription?: any
    yarnQueue?: any
    teamIdList?: any
    teamName: string
    flinkRestUrl?: any
    buildStatus: number
    appControl: AppControl
    fsOperator: any
    workspace: any
    k8sPodTemplates: {
        empty: boolean
    }
    streamParkJob: boolean
    hadoopUser: string
}

interface AppControl {
    allowStart: boolean
    allowStop: boolean
    allowBuild: boolean
    allowView?: boolean
}